+++ src/main/java/br/edu/unifio/odonto/desapego/user/dto/UserMeResponse.java
@@ -1,0 +1,20 @@
+package br.edu.unifio.odonto.desapego.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserMeResponse {
+
+    private UUID id;
+    private String email;
+    private String name;
+    private String role;
+}